change struct's name and impl Display
- change struct `MacAddress` -> `MacAddr`
- impl fmt() for MacAddr
