+++ src/address/mod.rs
@@ -1,1 +1,1 @@
-pub mod mac_address;
+pub mod mac_addr;
