+++ src/datalink/ethernet/mod.rs
@@ -21,7 +21,7 @@ pub struct EthernetFrame<'a> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::address::mac_address::MacAddress;
+    use crate::address::mac_addr::MacAddr;
     use header::EthernetHeader;
 
     #[test]
@@ -39,8 +39,8 @@ mod tests {
         let ethernet_frame = EthernetFrame::from_bytes(bytes);
 
         let expect_header = EthernetHeader {
-            destination_mac_address: MacAddress(0xff, 0xff, 0xff, 0xff, 0xff, 0xff),
-            source_mac_address: MacAddress(0x00, 0x00, 0x00, 0x00, 0x00, 0x00),
+            destination_mac_address: MacAddr(0xff, 0xff, 0xff, 0xff, 0xff, 0xff),
+            source_mac_address: MacAddr(0x00, 0x00, 0x00, 0x00, 0x00, 0x00),
             ethertype: [0x08, 0x00],
         };
         let expect_payload: &[u8] = &[
