+++ src/lib.rs
@@ -3,7 +3,7 @@
 pub mod address;
 pub mod datalink;
 
-use crate::address::mac_address::MacAddress;
+use crate::address::mac_addr::MacAddr;
 use datalink::ethernet::EthernetFrame;
 use datalink::ethernet::header::EthernetHeader;
 
@@ -22,8 +22,8 @@ impl PacketBuilder {
 
     pub fn ethernet<'a>(
         &self,
-        destination: MacAddress,
-        source: MacAddress,
+        destination: MacAddr,
+        source: MacAddr,
         ethertype: [u8; 2],
         payload: &'a [u8],
     ) -> EthernetFrame<'a> {
@@ -53,15 +53,15 @@ mod tests {
         let builder = PacketBuilder::new();
         let expect = EthernetFrame {
             header: EthernetHeader {
-                destination_mac_address: MacAddress(0xff, 0xff, 0xff, 0xff, 0xff, 0xff),
-                source_mac_address: MacAddress(0x00, 0x00, 0x00, 0x00, 0x00, 0x00),
+                destination_mac_address: MacAddr(0xff, 0xff, 0xff, 0xff, 0xff, 0xff),
+                source_mac_address: MacAddr(0x00, 0x00, 0x00, 0x00, 0x00, 0x00),
                 ethertype: [0x08, 0x00],
             },
             payload: &[0x45, 0x00],
         };
         let actual = builder.ethernet(
-            MacAddress(0xff, 0xff, 0xff, 0xff, 0xff, 0xff),
-            MacAddress(0x00, 0x00, 0x00, 0x00, 0x00, 0x00),
+            MacAddr(0xff, 0xff, 0xff, 0xff, 0xff, 0xff),
+            MacAddr(0x00, 0x00, 0x00, 0x00, 0x00, 0x00),
             [0x08, 0x00],
             &[0x45, 0x00],
         );
